Move functions from command_template
