+++ src/converters.rs
@@ -351,3 +351,79 @@ pub fn u256_to_bytes32_little(x: &U256) -> [u8; 32] {
     x.to_little_endian(&mut bytes);
     bytes
 }
+
+/// Converts an unsigned integer to a decimal string representation.
+///
+/// # Arguments
+///
+/// * `uint` - The unsigned integer to convert.
+/// * `decimal` - The number of decimal places to use.
+///
+/// # Returns
+///
+/// A string representation of the decimal value.
+pub fn uint_to_decimal_string(uint: u128, decimal: usize) -> String {
+    // Convert amount to string in wei format (no decimals)
+    let uint_str = uint.to_string();
+    let uint_length = uint_str.len();
+
+    // Create result vector with max length
+    // If less than 18 decimals, then 2 extra for "0.", otherwise one extra for "."
+    let mut result = vec![
+        '0';
+        if uint_length > decimal {
+            uint_length + 1
+        } else {
+            decimal + 2
+        }
+    ];
+    let result_length = result.len();
+
+    // Difference between result and amount array index when copying
+    // If more than 18, then 1 index diff for ".", otherwise actual diff in length
+    let mut delta = if uint_length > decimal {
+        1
+    } else {
+        result_length - uint_length
+    };
+
+    // Boolean to indicate if we found a non-zero digit when scanning from last to first index
+    let mut found_non_zero_decimal = false;
+
+    let mut actual_result_len = 0;
+
+    // In each iteration we fill one index of result array (starting from end)
+    for i in (0..result_length).rev() {
+        // Check if we have reached the index where we need to add decimal point
+        if i == result_length - decimal - 1 {
+            // No need to add "." if there was no value in decimal places
+            if found_non_zero_decimal {
+                result[i] = '.';
+                actual_result_len += 1;
+            }
+            // Set delta to 0, as we have already added decimal point (only for amount_length > 18)
+            delta = 0;
+        }
+        // If amountLength < 18 and we have copied everything, fill zeros
+        else if uint_length <= decimal && i < result_length - uint_length {
+            result[i] = '0';
+            actual_result_len += 1;
+        }
+        // If non-zero decimal is found, or decimal point inserted (delta == 0), copy from amount array
+        else if found_non_zero_decimal || delta == 0 {
+            result[i] = uint_str.chars().nth(i - delta).unwrap();
+            actual_result_len += 1;
+        }
+        // If we find non-zero decimal for the first time (trailing zeros are skipped)
+        else if uint_str.chars().nth(i - delta).unwrap() != '0' {
+            result[i] = uint_str.chars().nth(i - delta).unwrap();
+            actual_result_len += 1;
+            found_non_zero_decimal = true;
+        }
+    }
+
+    // Create final result string with correct length
+    let compact_result: String = result.into_iter().take(actual_result_len).collect();
+
+    compact_result
+}
